IVC: move constraints related functions
The "interpreter" design aims to split the witness and constraint environment.
This is a first step into the clean refactorisation of the IVC circuit.
